+++ src/server/material_tiers/index.js
@@ -165,57 +165,49 @@ const addAutomaticMaterialRecipes = (event) => {
 
 /**
  * Adjusts recipes relating to the deep storage units (super/quantum chests and tanks).
- * 
- * This has to be manually adjusted to fix the internal crate 
- * 
+ *
+ * This has to be manually adjusted to fix the internal crate
+ *
  * @param {Internal.RecipesEventJS} event
  */
 const adjustSuperChestRecipes = (event) => {
     event.remove({ output: /gtceu:.v_super_chest/ });
-    
-    event.shaped(
-        "gtceu:lv_super_chest",
-        ["CPC", "PKP", "CPC"],
-        {
+
+    event
+        .shaped("gtceu:lv_super_chest", ["CPC", "PKP", "CPC"], {
             P: GT_MACHINE_TIERS.LV.primaryPlate,
             C: GT_MACHINE_TIERS.LV.circuitTag,
             K: "gtceu:bronze_crate",
-        }
-    ).id("nijika:super_chests/lv");
+        })
+        .id("nijika:super_chests/lv");
 
     // lol at how completely wrongly coloured these aree
-    event.shaped(
-        "gtceu:mv_super_chest",
-        ["CPC", "PKP", "CPC"],
-        {
+    event
+        .shaped("gtceu:mv_super_chest", ["CPC", "PKP", "CPC"], {
             P: GT_MACHINE_TIERS.MV.primaryPlate,
             C: GT_MACHINE_TIERS.MV.circuitTag,
             K: "gtceu:steel_crate",
-        }
-    ).id("nijika:super_chests/mv");
+        })
+        .id("nijika:super_chests/mv");
 
-    event.shaped(
-        "gtceu:hv_super_chest",
-        ["CPC", "PKP", "CFC"],
-        {
+    event
+        .shaped("gtceu:hv_super_chest", ["CPC", "PKP", "CFC"], {
             P: GT_MACHINE_TIERS.HV.primaryPlate,
             C: GT_MACHINE_TIERS.HV.circuitTag,
             K: "gtceu:aluminium_crate",
-            F: "gtceu:lv_field_generator"
-        }
-    ).id("nijika:super_chests/hv");
+            F: "gtceu:lv_field_generator",
+        })
+        .id("nijika:super_chests/hv");
 
-    event.shaped(
-        "gtceu:ev_super_chest",
-        ["CPC", "PKP", "CFC"],
-        {
+    event
+        .shaped("gtceu:ev_super_chest", ["CPC", "PKP", "CFC"], {
             P: GT_MACHINE_TIERS.EV.primaryPlate,
             C: GT_MACHINE_TIERS.EV.circuitTag,
-            K: "gtceu:titanium_crate",  // TODO: nitinol
-            F: "gtceu:mv_field_generator"
-        }
-    ).id("nijika:super_chests/ev");
-}
+            K: "gtceu:titanium_crate", // TODO: nitinol
+            F: "gtceu:mv_field_generator",
+        })
+        .id("nijika:super_chests/ev");
+};
 
 /**
  * Adjusts recipes relating to the material system and BI's adjusted tiers.