+++ src/server/tier00.js
@@ -156,7 +156,7 @@ export const doTier00Content = (event) => {
             event.recipes.create.deploying("gtceu:resin_circuit_board", [
                 "gtceu:resin_circuit_board",
                 "gtceu:copper_single_wire",
-            ])
+            ]),
         ])
         .transitionalItem("gtceu:resin_circuit_board")
         .loops(3)